+++ trasitarios/src/Veiculos/Van.java
@@ -19,7 +19,7 @@ public class Van extends Veiculo{
         super.set_Refrigerado(true);
     }
     
-    public Van(String marca, String matricula, String gasolina, float autonomia, float custoKm, float desgaste, float carga,boolean refrigerado)
+    public Van(String marca, String matricula, String gasolina, double autonomia, double custoKm, double desgaste, double carga,boolean refrigerado)
     {
         super(marca,matricula,custoKm,desgaste,carga);
         super.set_Refrigerado(true);
@@ -36,11 +36,11 @@ public class Van extends Veiculo{
     @Override
     public String get_Matricula(){return super.get_Matricula();}
     @Override
-    public float get_Desgaste(){return super.get_Desgaste();}
+    public double get_Desgaste(){return super.get_Desgaste();}
     @Override
-    public float get_CustoKm(){return super.get_CustoKm();}
+    public double get_CustoKm(){return super.get_CustoKm();}
     @Override
-    public float get_Carga(){return super.get_Carga();}
+    public double get_Carga(){return super.get_Carga();}
     @Override
     public boolean get_Refrigerado(){return super.get_Refrigerado();}
        
@@ -50,11 +50,11 @@ public class Van extends Veiculo{
     @Override
     public void set_Matricula(String matricula){super.set_Matricula(matricula);}
     @Override
-    public void set_CustoKm(float custoKm){super.set_CustoKm(custoKm);}
+    public void set_CustoKm(double custoKm){super.set_CustoKm(custoKm);}
     @Override
-    public void set_Desgaste(float desgaste){super.set_Desgaste(desgaste);}
+    public void set_Desgaste(double desgaste){super.set_Desgaste(desgaste);}
     @Override
-    public void set_Carga(float carga){super.set_Carga(carga);}
+    public void set_Carga(double carga){super.set_Carga(carga);}
     @Override
     public void set_Refrigerado(boolean refrigeracao){super.set_Refrigerado(true);}
